Removed old PetMode, Removed Areas with Tank check
Removed old PetMode and all new features rely on the cycle mode.

Removed areas that check for tank, but if you have the attachment and you are not a tank for some reason then it should be able to display and track.

Removed an enable ear1 which no longer needed.

Completed removed self command for changing modes. This is handle entirely by cycle now. No longer safe to utilize this command.