Issue with Precast, Better logic to pet tp
In the event of a weaponskill being used and pet is not out the gearswap would error out. This has be corrected.

Pet Weaponskill set will now only activate when the player is not engaged. If player is engaged the Pet Weaponskill set is already handled in the player aftercast.