Update Weaponskill for Pet, Update README
Adjusted to prevent the equipping prior to pet weaponskill if the player does have enough TP. This will also be overridden by spam and DD in the event you want this to overcome player.

Corrected Format for the README, added explanations of how the Pet Weaponskill Functions and how the Enmity Gear is decided.

Added to Predict for when the abilities activate or deus are used.