WARNING - EFFECTS SETS
This version adds in comments to help newer players understand what modes to be in for each set to activate.

This adds a Hybrid Mode option for Ranged for Pets.

This also takes away the old sets.idle.PetDT and has been replaced by sets.pet.EmergencyDT.

Also, big clean up to weapon skill sets for the puppet.