Code Clean Up/Minor Feature/Minor Bug Fix
Cleaned up unneeded code and comments. Added in comments above the various states of how to cycle/toggle these with commands or keybinds.

Adjusted the Main HUB window for hiding just like the other various states. Added in the option to now hide the options portion of the HUB.

Added in the ability to set up a custom gear lock that can be toggled via in game command.

Added in the MagicalDefenseMode for PetMDT in case players have gear more specific to magic.

Minor bug fix for equipping pet weapon skill gear prior to weapon skill. This is now 100% only activated when the player is idle and pet is fighting.