Updates for UI/Bug Fix/New OffenseMode
Update to UI to include a Lite Mode. As well as some coloring adding for when an option is off or on.
--Courtesy of Tuna

Bug fix for the WS Shinjin Spiral, missing brackets.
--Courtesy of Tuna

Update to OffenseMode by adding a "Trusts" mode, this mode enables you to be engaged, but favors idle.Pet sets for when you want trusts to fight, but not be near mob. 
--Courtesy of Tuna