[Bug] Endless Maneuver Queue
The queue for maneuvers could enter an endless loop causing it to recast the same lost maneuvers over and over when it should ignore adding a new one if we have over 3.

Added the current total of queue on the window under Pet Skills.

Added a new command/keybind (=) to allow players to manually reset the queue.